Refactor and updating versions
