+++ .editorconfig
@@ -1,22 +1,24 @@
-# EditorConfig helps developers define and maintain consistent
-# coding styles between different editors and IDEs
-# editorconfig.org
 
 root = true
 
 [*]
-max_line_length = 160
+charset = utf-8
 indent_style = space
 indent_size = 4
-end_of_line = lf
-charset = utf-8
-# A file must end with an empty line - this is good for version control systems
 insert_final_newline = true
+end_of_line = lf
+max_line_length = 160
 trim_trailing_whitespace = true
 
-[*.{diff, md}]
-trim_trailing_whitespace = false
+[{*.yml,*.yaml}]
+indent_size = 2
 
-[Makefile]
+[{*.json,*.xml,*.golden}]
+insert_final_newline = false
+
+[*.sh]
+indent_size = 2
+
+[{Makefile,go.mod,go.sum,*.go}]
 indent_style = tab
-tab_width = 4
+indent_size = 8
