+++ go.mod
@@ -1,11 +1,21 @@
 module github.com/arnobroekhof/rotate-writer
 
-go 1.14
+go 1.15
 
 require (
-	github.com/Azure/azure-storage-blob-go v0.8.0
-	github.com/Azure/go-autorest/autorest/adal v0.8.2 // indirect
-	github.com/minio/minio-go/v6 v6.0.50
-	github.com/stretchr/testify v1.5.1
-	github.com/testcontainers/testcontainers-go v0.3.1
+	github.com/Azure/azure-storage-blob-go v0.12.0
+	github.com/google/uuid v1.1.4 // indirect
+	github.com/json-iterator/go v1.1.10 // indirect
+	github.com/klauspost/cpuid v1.3.1 // indirect
+	github.com/minio/md5-simd v1.1.1 // indirect
+	github.com/minio/minio-go/v6 v6.0.57
+	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
+	github.com/modern-go/reflect2 v1.0.1 // indirect
+	github.com/stretchr/testify v1.6.1
+	github.com/testcontainers/testcontainers-go v0.9.0
+	golang.org/x/crypto v0.0.0-20201221181555-eec23a3978ad // indirect
+	golang.org/x/net v0.0.0-20201224014010-6772e930b67b // indirect
+	golang.org/x/sys v0.0.0-20210104204734-6f8348627aad // indirect
+	golang.org/x/text v0.3.4 // indirect
+	gopkg.in/ini.v1 v1.62.0 // indirect
 )
