+++ pkg/azure/blob_writer_test.go
